V1.3.x 依赖优化

+++ joggle-server/src/main/java/com/wuweibi/bullet/orders/service/OrdersRefundService.java
@@ -23,5 +23,5 @@ public interface OrdersRefundService extends IService<OrdersRefund> {
      */
     Page<OrdersRefundVO> getPage(Page pageInfo, OrdersRefundParam params);
 
-    int countStatusByOrderId(Long orderId, int status);
+    Long countStatusByOrderId(Long orderId, int status);
 }
